v1.3.1.2 update

+++ src/serialize.h
@@ -114,7 +114,7 @@ inline unsigned int GetSerializeSize(signed int a,         int, int=0) { return 
 inline unsigned int GetSerializeSize(unsigned int a,       int, int=0) { return sizeof(a); }
 inline unsigned int GetSerializeSize(signed long a,        int, int=0) { return sizeof(a); }
 inline unsigned int GetSerializeSize(unsigned long a,      int, int=0) { return sizeof(a); }
-inline unsigned int GetSerializeSize(signed long long a,   int, int=0) { return sizeof(a); }
+inline unsigned int GetSerializeSize(__int128 a,           int, int=0) { return sizeof(a); }
 inline unsigned int GetSerializeSize(unsigned long long a, int, int=0) { return sizeof(a); }
 inline unsigned int GetSerializeSize(float a,              int, int=0) { return sizeof(a); }
 inline unsigned int GetSerializeSize(double a,             int, int=0) { return sizeof(a); }
@@ -130,7 +130,7 @@ template<typename Stream> inline void Serialize(Stream& s, signed long a,       
 template<typename Stream> inline void Serialize(Stream& s, unsigned long a,      int, int=0) { WRITEDATA(s, a); }
 template<typename Stream> inline void Serialize(Stream& s, signed long long a,   int, int=0) { WRITEDATA(s, a); }
 template<typename Stream> inline void Serialize(Stream& s, unsigned long long a, int, int=0) { WRITEDATA(s, a); }
-template<typename Stream> inline void Serialize(Stream& s, float a,              int, int=0) { WRITEDATA(s, a); }
+template<typename Stream> inline void Serialize(Stream& s, __int128 a,           int, int=0) { WRITEDATA(s, a); }
 template<typename Stream> inline void Serialize(Stream& s, double a,             int, int=0) { WRITEDATA(s, a); }
 
 template<typename Stream> inline void Unserialize(Stream& s, char& a,               int, int=0) { READDATA(s, a); }
@@ -145,7 +145,7 @@ template<typename Stream> inline void Unserialize(Stream& s, unsigned long& a,  
 template<typename Stream> inline void Unserialize(Stream& s, signed long long& a,   int, int=0) { READDATA(s, a); }
 template<typename Stream> inline void Unserialize(Stream& s, unsigned long long& a, int, int=0) { READDATA(s, a); }
 template<typename Stream> inline void Unserialize(Stream& s, float& a,              int, int=0) { READDATA(s, a); }
-template<typename Stream> inline void Unserialize(Stream& s, double& a,             int, int=0) { READDATA(s, a); }
+template<typename Stream> inline void Unserialize(Stream& s, __int128& a,           int, int=0) { READDATA(s, a); }
 
 inline unsigned int GetSerializeSize(bool a, int, int=0)                          { return sizeof(char); }
 template<typename Stream> inline void Serialize(Stream& s, bool a, int, int=0)    { char f=a; WRITEDATA(s, f); }
@@ -935,7 +935,7 @@ public:
     {
         // Read from the beginning of the buffer
         assert(nSize >= 0);
-        unsigned int nReadPosNext = nReadPos + nSize;
+        uint64_t nReadPosNext = nReadPos + nSize;
         if (nReadPosNext >= vch.size())
         {
             if (nReadPosNext > vch.size())
